better approach by spliting logic in diferent components for the textarea with variables or the minimal variant
